email-sender/requirements.txt
python-decouple==3.4
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.21
Flask-Migrate==3.0.1
Flask-Script==2.0.6
Flask-Cors==3.0.10
requests==2.26.0
mysqlclient==2.0.2
pika==1.2.0
Flask-WTF==0.15.1
Flask-Login==0.5.0
Flask-OAuthlib==0.9.6
google-api-python-client==2.15.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4